requirements.txt
tabulate
matplotlib
scipy
click
numpy
python-dateutil
requests
